Finalized state machine structure
Still need to test and implement backup/restore at the states that
implement them.
Not quite sure how addresses/ranks should be represented later on.
Should HashLinks contain a postgres member? Or should that be added on
in another crate?

Simple wiki currently takes about 11 minutes, of which ~10 minutes are
for pagerank. Maybe it can be parallelized