Completed to/from file restoring/loading and switch to FnvHashMap
Saving/restoring has also been tested with the simplewiki. Rank data
will be in a different order if loaded/restored because its internal rep
is a hashmap, but when re-sorted they're identical again.
Still need to figure out rank representation for HashLinks. I'm leaning
for a (postgre) db member, and then HashLinks can have a bunch of
requests built in, e.g. bfs and rank lookups.