Adding diesel functionality before reorganizing project
I think it will all be in one crate. It might be interesting to make a
'links' crate and a 'site' crate that uses it, but I don't think that
has any major advantages. The projects aren't really able to stand on
their own, and compile time isn't an issue, so I probably won't separate
them. Instead I'll probably put all the link_state stuff in its own
module, the cache stuff in a sibling module, and have lib.rs import them
and handle the Rocket stuff