Tidy up code; delete unused blocks and use clippy
Things are mostly tidied up. The parsing step is kind of a mess, doing a
fair amount of redundant work (that is maybe still not complete? it
throws warnings at runtime...) and inheriting the dumps' lack of
organization (about a third of the project code is in one file). But it
seems to work, and at the very least seems to err on the side of
deleting valid data rather than relying on invalid data.
The next step is probably performance profiling. Running 4 procs to make
continuous Random->Random requests pegged the cpus but other requests
would still go through, which is a good sign. They also bumped the db
size up to ~500MB, which is funny. Maybe there should be a lower limit
on what searches are cached. At some point it becomes cheaper to
recompute a short path rather than store it in memory. Maybe with some
analytics that decision can be made (e.g. if 90% of searches complete in
<4 iterations then we can probably save a lot of space).