Begin benchmarking major components of searching
Want to reduce unnecessary overhead if this is actually deployed; the
main ways this might be most effective are by reducing unnecessary
database queries, which can't really compare speed-wise with in-memory
requests.

Maintaining an historical cache as a database and querying it / sorting
results could be replaced with maintaining 2 or 3 historical "cache"s in
shared memory and updating them as necessary. Cache db lookups range
from 200-260 microseconds for the simple wiki, which seems orders of
magnitude longer than necessary.

Looking up an article id by its title in a database requires about 40
microseconds on Simple wiki; this isn't terrible compared with some
other components of searching, but it is run twice per request every
time. It might (?) be faster to maintain an in-memory lookup table
mapping strings to page_ids. This is only feasible if
a) this doesn't put the memory footprint above/near 8GB: this will
require swapping on most machines, which is almost certainly not worth
it
b) using the default Rust hasher doesn't slow things down too much; fnv
seems like a bad idea for this case; maybe there's another fast option?
c) trashing the shitty fuzzy searching is not a big deal; maybe the
proper way to do real fuzzy searching can be done with a hash table?

It is also worth looking into conditionally caching searches; shorter
breadth-first searches are much faster than a db query, and storing
every search might be unnecessary; does postgres slow down as the table
gets bigger (probably not)? is a marginal slowdown worth a major hard
drive memory saving? it's worth looking into.