Rework rank data importing/exporting
Ranks aren't actually necessary for any of the bfs stuff, and they're
the most expensive part to compute, so they're now optional.
Instead, it is mandatory to create a lookup table of titles which
contain capitalized forms of any titles which have unique
capitalizations. Unfortunately there are some weird corner cases, like
when the capitalized form of one word (e.g. `Gnu`) is another article
(`GNU`); so all entries have to be annotated.
Next step is to see how expensive that is and maybe replace 2 of the 3
mandatory db lookups with 2 hash table lookups instead in a way that is
somewhat case-insensitive