Begin reworking history mechanism
Looking up the history from a db by arbitrary fields is suboptimal.
Instead it is straightforward to only maintain a few independent
collections of searches that greedily determine whether to update. This
makes it impossible to sort by "random", but that was already pretty
useless because it's basically sorting by "recent". However it is now
also impossible to sort by "popular". I guess it's not a big deal.

This caching system is easy to make but hard to make fast. They need to
be quick to update, copy, and sort. They also have to be thread-safe.

Recent searches are enclosed in a mutex because they'll be updated
pretty much every time. Long searches searches are in a RwLock because
they are updated rarely but requested frequently (maybe make that the
default?)