Switch StackCache to not use references everywhere
I think it might be impossible to use references how I was trying to in
a shared concurrent safe way. For now I'm just going to get everything
working with unnecessary uses of `clone()`s, and it can be refactored
later. This shouldn't matter too much, because it should still be orders
of magnitude faster than a database request