CacheStack half-implemented, cleanups
Gonna have to change cachestack; instead of a struct w/ fields, each
field (LengthStack and RecentStack) will be separate and wrapped in a
lock (rwlock and mutex) and an in an Arc and then passed around. This
will work better with rocket.
Unless atomics would be better. They're harder to optimize and maybe
harder to reason about, but the alternative is to clone() every time a
mutex/rwlock is locked, which isn't great. It's a lot better than the
alternative, but still not great

EDIT: on the other hand...
Atomics are hard. It might be possible to use a combination of
crossbeam's atomic stack and evmap's set, perhaps requiring 3 different
stacks (src, dst, len, if there's no good way to atomically compose
atomics).
This would be a little clunky, and would mean the new_cache and
long_cache are implemented totally differently.
Also, it might not even help. We want to wrap article titles inside
atomic ptrs, but it looks like Rocket route functions require params to
have only static lifetimes, which would destroy most of the purpose of
using atomics in the first place.
For now, I'm going to just use a normal locking structure and maybe
drastically refactor later on.

¯\_(ツ)_/¯