Fix cache/history ordering order
Both the 'recent' and 'longest' caches now store the correct values and
display them in the right order. This incurs a small performance hit,
but not a notable one.

What's next: not sure;
Could refactor and make sql entry much smaller (omit count and
timestamp), or could populate initial stack cache from what's stored in
the database.
I'll probably do the latter. It has small accuracy hits, though: the
recency won't be updated if a search is re-performed (shouldn't be?) and
it won't include recent short searches (should it?)