Need to amend; title to page_id must be composed in the link_data state
So the titles db has to be built from the beginning.
While we're at it, remove the metadata dump because it's no longer
necessary. Tack on the title data to the original links dump and don't
even represent pagerank. Then replace the RankData state with an
AnalData state (or something with a better name), where analysis steps
can be run from the command line, if necessary.