take a shot every time you underestimate heap alloc times
this is actually like strictly faster than before. but the starting ihm
size needs to be lower (was 1024, is now 64, stdlib's is 32). apparently
like 5 allocs were causing a 500% slowdown on small searches. but now
with a 50% resize threshold it's about 50% faster (!). just need some
profiling to verify it's not using way more memory than stdlib. is seems
like it isn't, but the hashmap src is a little funky and I'm not sure
I'm reading it right

EDIT: bfs w/ ihm uses less memory also. looks to be about 5% less, but
it's still definitely less with a 50% resize threshold

and for some reason jemalloc seems to segfault in valgrind :/
but system alloc works