only use one `Entry` type
Previously we had both `Entry` and `IndexedEntry`, which were the same
except the latter had a `page_id` associated with it. If we're storing a
table of entries and referring to each by its index within the table,
then there'll never be a situation in which we don't need the `page_id`.
So get rid of `Entry`, rename `IndexedEntry` to it, and fix the errors